Create a new MessageEditor
