+++ crates/agent_ui/src/acp/thread_view.rs
@@ -10,10 +10,7 @@ use agent_settings::{AgentSettings, NotifyWhenAgentWaiting};
 use audio::{Audio, Sound};
 use buffer_diff::BufferDiff;
 use collections::{HashMap, HashSet};
-use editor::{
-    AnchorRangeExt, ContextMenuOptions, ContextMenuPlacement, Editor, EditorElement, EditorMode,
-    EditorStyle, MinimapVisibility, MultiBuffer, PathKey,
-};
+use editor::{Editor, EditorMode, MinimapVisibility, MultiBuffer, PathKey};
 use file_icons::FileIcons;
 use gpui::{
     Action, Animation, AnimationExt, App, BorderStyle, EdgesRefinement, Empty, Entity, EntityId,
@@ -22,20 +19,15 @@ use gpui::{
     Transformation, UnderlineStyle, WeakEntity, Window, WindowHandle, div, linear_color_stop,
     linear_gradient, list, percentage, point, prelude::*, pulsating_between,
 };
+use language::Buffer;
 use language::language_settings::SoftWrap;
-use language::{Buffer, Language};
 use markdown::{HeadingLevelStyles, Markdown, MarkdownElement, MarkdownStyle};
-use parking_lot::Mutex;
-use project::{CompletionIntent, Project};
+use project::Project;
 use rope::Point;
 use settings::{Settings as _, SettingsStore};
-use std::path::PathBuf;
-use std::{
-    cell::RefCell, collections::BTreeMap, path::Path, process::ExitStatus, rc::Rc, sync::Arc,
-    time::Duration,
-};
+use std::{cell::RefCell, collections::BTreeMap, process::ExitStatus, rc::Rc, time::Duration};
 use terminal_view::TerminalView;
-use text::{Anchor, BufferSnapshot};
+use text::Anchor;
 use theme::ThemeSettings;
 use ui::{
     Disclosure, Divider, DividerColor, KeyBinding, PopoverMenuHandle, Scrollbar, ScrollbarState,
@@ -43,13 +35,12 @@ use ui::{
 };
 use util::{ResultExt, size::format_file_size, time::duration_alt_display};
 use workspace::{CollaboratorId, Workspace};
-use zed_actions::agent::{Chat, NextHistoryMessage, PreviousHistoryMessage, ToggleModelSelector};
+use zed_actions::agent::{Chat, ToggleModelSelector};
 
 use crate::acp::AcpModelSelectorPopover;
-use crate::acp::completion_provider::{ContextPickerCompletionProvider, MentionSet};
+use crate::acp::message_editor::{MessageEditor, MessageEditorEvent};
 use crate::acp::message_history::MessageHistory;
 use crate::agent_diff::AgentDiff;
-use crate::message_editor::{MAX_EDITOR_LINES, MIN_EDITOR_LINES};
 use crate::ui::{AgentNotification, AgentNotificationEvent};
 use crate::{
     AgentDiffPane, AgentPanel, ExpandMessageEditor, Follow, KeepAll, OpenAgentDiff, RejectAll,
@@ -64,11 +55,8 @@ pub struct AcpThreadView {
     thread_state: ThreadState,
     diff_editors: HashMap<EntityId, Entity<Editor>>,
     terminal_views: HashMap<EntityId, Entity<TerminalView>>,
-    message_editor: Entity<Editor>,
+    message_editor: Entity<MessageEditor>,
     model_selector: Option<Entity<AcpModelSelectorPopover>>,
-    message_set_from_history: Option<BufferSnapshot>,
-    _message_editor_subscription: Subscription,
-    mention_set: Arc<Mutex<MentionSet>>,
     notifications: Vec<WindowHandle<AgentNotification>>,
     notification_subscriptions: HashMap<WindowHandle<AgentNotification>, Vec<Subscription>>,
     last_error: Option<Entity<Markdown>>,
@@ -83,7 +71,7 @@ pub struct AcpThreadView {
     terminal_expanded: bool,
     message_history: Rc<RefCell<MessageHistory<Vec<acp::ContentBlock>>>>,
     _cancel_task: Option<Task<()>>,
-    _subscriptions: [Subscription; 1],
+    _subscriptions: [Subscription; 2],
 }
 
 enum ThreadState {
@@ -109,80 +97,25 @@ impl AcpThreadView {
         workspace: WeakEntity<Workspace>,
         project: Entity<Project>,
         message_history: Rc<RefCell<MessageHistory<Vec<acp::ContentBlock>>>>,
-        min_lines: usize,
-        max_lines: Option<usize>,
         window: &mut Window,
         cx: &mut Context<Self>,
     ) -> Self {
-        let language = Language::new(
-            language::LanguageConfig {
-                completion_query_characters: HashSet::from_iter(['.', '-', '_', '@']),
-                ..Default::default()
-            },
-            None,
-        );
-
-        let mention_set = Arc::new(Mutex::new(MentionSet::default()));
-
         let message_editor = cx.new(|cx| {
-            let buffer = cx.new(|cx| Buffer::local("", cx).with_language(Arc::new(language), cx));
-            let buffer = cx.new(|cx| MultiBuffer::singleton(buffer, cx));
-
-            let mut editor = Editor::new(
-                editor::EditorMode::AutoHeight {
-                    min_lines,
-                    max_lines: max_lines,
-                },
-                buffer,
-                None,
+            MessageEditor::new(
+                workspace.clone(),
+                project.clone(),
+                message_history.clone(),
                 window,
                 cx,
-            );
-            editor.set_placeholder_text("Message the agent － @ to include files", cx);
-            editor.set_show_indent_guides(false, cx);
-            editor.set_soft_wrap();
-            editor.set_use_modal_editing(true);
-            editor.set_completion_provider(Some(Rc::new(ContextPickerCompletionProvider::new(
-                mention_set.clone(),
-                workspace.clone(),
-                cx.weak_entity(),
-            ))));
-            editor.set_context_menu_options(ContextMenuOptions {
-                min_entries_visible: 12,
-                max_entries_visible: 12,
-                placement: Some(ContextMenuPlacement::Above),
-            });
-            editor
+            )
         });
-
-        let message_editor_subscription =
-            cx.subscribe(&message_editor, |this, editor, event, cx| {
-                if let editor::EditorEvent::BufferEdited = &event {
-                    let buffer = editor
-                        .read(cx)
-                        .buffer()
-                        .read(cx)
-                        .as_singleton()
-                        .unwrap()
-                        .read(cx)
-                        .snapshot();
-                    if let Some(message) = this.message_set_from_history.clone()
-                        && message.version() != buffer.version()
-                    {
-                        this.message_set_from_history = None;
-                    }
-
-                    if this.message_set_from_history.is_none() {
-                        this.message_history.borrow_mut().reset_position();
-                    }
-                }
-            });
-
-        let mention_set = mention_set.clone();
 
         let list_state = ListState::new(0, gpui::ListAlignment::Bottom, px(2048.0));
 
-        let subscription = cx.observe_global_in::<SettingsStore>(window, Self::settings_changed);
+        let subscriptions = [
+            cx.observe_global_in::<SettingsStore>(window, Self::settings_changed),
+            cx.subscribe_in(&message_editor, window, Self::on_message_editor_event),
+        ];
 
         Self {
             agent: agent.clone(),
@@ -191,9 +124,6 @@ impl AcpThreadView {
             thread_state: Self::initial_state(agent, workspace, project, window, cx),
             message_editor,
             model_selector: None,
-            message_set_from_history: None,
-            _message_editor_subscription: message_editor_subscription,
-            mention_set,
             notifications: Vec::new(),
             notification_subscriptions: HashMap::default(),
             diff_editors: Default::default(),
@@ -209,7 +139,7 @@ impl AcpThreadView {
             editor_expanded: false,
             terminal_expanded: true,
             message_history,
-            _subscriptions: [subscription],
+            _subscriptions: subscriptions,
             _cancel_task: None,
         }
     }
@@ -377,187 +307,64 @@ impl AcpThreadView {
 
     fn set_editor_is_expanded(&mut self, is_expanded: bool, cx: &mut Context<Self>) {
         self.editor_expanded = is_expanded;
-        self.message_editor.update(cx, |editor, _| {
-            if self.editor_expanded {
-                editor.set_mode(EditorMode::Full {
-                    scale_ui_elements_with_buffer_font_size: false,
-                    show_active_line_background: false,
-                    sized_by_content: false,
-                })
-            } else {
-                editor.set_mode(EditorMode::AutoHeight {
-                    min_lines: MIN_EDITOR_LINES,
-                    max_lines: Some(MAX_EDITOR_LINES),
-                })
-            }
+        self.message_editor.update(cx, |editor, cx| {
+            editor.set_expanded(is_expanded, cx);
         });
         cx.notify();
     }
 
-    fn chat(&mut self, _: &Chat, window: &mut Window, cx: &mut Context<Self>) {
+    pub fn on_message_editor_event(
+        &mut self,
+        _: &Entity<MessageEditor>,
+        event: &MessageEditorEvent,
+        window: &mut Window,
+        cx: &mut Context<Self>,
+    ) {
+        match event {
+            MessageEditorEvent::Chat => self.chat(window, cx),
+        }
+    }
+
+    fn chat(&mut self, window: &mut Window, cx: &mut Context<Self>) {
         self.last_error.take();
 
-        let mut ix = 0;
-        let mut chunks: Vec<acp::ContentBlock> = Vec::new();
-        let project = self.project.clone();
+        let Some(thread) = self.thread().cloned() else {
+            return;
+        };
 
-        let contents = self.mention_set.lock().contents(project, cx);
+        let contents = self
+            .message_editor
+            .update(cx, |message_editor, cx| message_editor.contents(cx));
+        let task = cx.spawn_in(window, async move |this, cx| {
+            let contents = contents.await?;
 
-        cx.spawn_in(window, async move |this, cx| {
-            let contents = match contents.await {
-                Ok(contents) => contents,
-                Err(e) => {
-                    this.update(cx, |this, cx| {
-                        this.last_error =
-                            Some(cx.new(|cx| Markdown::new(e.to_string().into(), None, None, cx)));
-                    })
-                    .ok();
-                    return;
-                }
-            };
+            if contents.is_empty() {
+                return Ok(());
+            }
 
             this.update_in(cx, |this, window, cx| {
-                this.message_editor.update(cx, |editor, cx| {
-                    let text = editor.text(cx);
-                    editor.display_map.update(cx, |map, cx| {
-                        let snapshot = map.snapshot(cx);
-                        for (crease_id, crease) in snapshot.crease_snapshot.creases() {
-                            // Skip creases that have been edited out of the message buffer.
-                            if !crease.range().start.is_valid(&snapshot.buffer_snapshot) {
-                                continue;
-                            }
-
-                            if let Some(mention) = contents.get(&crease_id) {
-                                let crease_range =
-                                    crease.range().to_offset(&snapshot.buffer_snapshot);
-                                if crease_range.start > ix {
-                                    chunks.push(text[ix..crease_range.start].into());
-                                }
-                                chunks.push(acp::ContentBlock::Resource(acp::EmbeddedResource {
-                                    annotations: None,
-                                    resource: acp::EmbeddedResourceResource::TextResourceContents(
-                                        acp::TextResourceContents {
-                                            mime_type: None,
-                                            text: mention.content.clone(),
-                                            uri: mention.uri.to_uri(),
-                                        },
-                                    ),
-                                }));
-                                ix = crease_range.end;
-                            }
-                        }
-
-                        if ix < text.len() {
-                            let last_chunk = text[ix..].trim_end();
-                            if !last_chunk.is_empty() {
-                                chunks.push(last_chunk.into());
-                            }
-                        }
-                    })
-                });
-
-                if chunks.is_empty() {
-                    return;
-                }
-
-                let Some(thread) = this.thread() else {
-                    return;
-                };
-                let task = thread.update(cx, |thread, cx| thread.send(chunks.clone(), cx));
-
-                cx.spawn(async move |this, cx| {
-                    let result = task.await;
-
-                    this.update(cx, |this, cx| {
-                        if let Err(err) = result {
-                            this.last_error =
-                                Some(cx.new(|cx| {
-                                    Markdown::new(err.to_string().into(), None, None, cx)
-                                }))
-                        }
-                    })
-                })
-                .detach();
-
-                let mention_set = this.mention_set.clone();
-
                 this.set_editor_is_expanded(false, cx);
-
-                this.message_editor.update(cx, |editor, cx| {
-                    editor.clear(window, cx);
-                    editor.remove_creases(mention_set.lock().drain(), cx)
-                });
-
                 this.scroll_to_bottom(cx);
+                this.message_editor.update(cx, |message_editor, cx| {
+                    message_editor.clear(window, cx);
+                });
+                this.message_history.borrow_mut().push(contents.clone());
+            })?;
+            let send = thread.update(cx, |thread, cx| thread.send(contents, cx))?;
+            send.await
+        });
 
-                this.message_history.borrow_mut().push(chunks);
-            })
-            .ok();
+        cx.spawn(async move |this, cx| {
+            if let Err(e) = task.await {
+                this.update(cx, |this, cx| {
+                    this.last_error =
+                        Some(cx.new(|cx| Markdown::new(e.to_string().into(), None, None, cx)));
+                    cx.notify()
+                })
+                .ok();
+            }
         })
         .detach();
-    }
-
-    fn previous_history_message(
-        &mut self,
-        _: &PreviousHistoryMessage,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if self.message_set_from_history.is_none() && !self.message_editor.read(cx).is_empty(cx) {
-            self.message_editor.update(cx, |editor, cx| {
-                editor.move_up(&Default::default(), window, cx);
-            });
-            return;
-        }
-
-        self.message_set_from_history = Self::set_draft_message(
-            self.message_editor.clone(),
-            self.mention_set.clone(),
-            self.project.clone(),
-            self.message_history
-                .borrow_mut()
-                .prev()
-                .map(|blocks| blocks.as_slice()),
-            window,
-            cx,
-        );
-    }
-
-    fn next_history_message(
-        &mut self,
-        _: &NextHistoryMessage,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if self.message_set_from_history.is_none() {
-            self.message_editor.update(cx, |editor, cx| {
-                editor.move_down(&Default::default(), window, cx);
-            });
-            return;
-        }
-
-        let mut message_history = self.message_history.borrow_mut();
-        let next_history = message_history.next();
-
-        let set_draft_message = Self::set_draft_message(
-            self.message_editor.clone(),
-            self.mention_set.clone(),
-            self.project.clone(),
-            Some(
-                next_history
-                    .map(|blocks| blocks.as_slice())
-                    .unwrap_or_else(|| &[]),
-            ),
-            window,
-            cx,
-        );
-        // If we reset the text to an empty string because we ran out of history,
-        // we don't want to mark it as coming from the history
-        self.message_set_from_history = if next_history.is_some() {
-            set_draft_message
-        } else {
-            None
-        };
     }
 
     fn open_agent_diff(&mut self, _: &OpenAgentDiff, window: &mut Window, cx: &mut Context<Self>) {
@@ -585,91 +392,6 @@ impl AcpThreadView {
         diff.update(cx, |diff, cx| {
             diff.move_to_path(PathKey::for_buffer(&buffer, cx), window, cx)
         })
-    }
-
-    fn set_draft_message(
-        message_editor: Entity<Editor>,
-        mention_set: Arc<Mutex<MentionSet>>,
-        project: Entity<Project>,
-        message: Option<&[acp::ContentBlock]>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Option<BufferSnapshot> {
-        cx.notify();
-
-        let message = message?;
-
-        let mut text = String::new();
-        let mut mentions = Vec::new();
-
-        for chunk in message {
-            match chunk {
-                acp::ContentBlock::Text(text_content) => {
-                    text.push_str(&text_content.text);
-                }
-                acp::ContentBlock::Resource(acp::EmbeddedResource {
-                    resource: acp::EmbeddedResourceResource::TextResourceContents(resource),
-                    ..
-                }) => {
-                    let path = PathBuf::from(&resource.uri);
-                    let project_path = project.read(cx).project_path_for_absolute_path(&path, cx);
-                    let start = text.len();
-                    let content = MentionUri::File(path).to_uri();
-                    text.push_str(&content);
-                    let end = text.len();
-                    if let Some(project_path) = project_path {
-                        let filename: SharedString = project_path
-                            .path
-                            .file_name()
-                            .unwrap_or_default()
-                            .to_string_lossy()
-                            .to_string()
-                            .into();
-                        mentions.push((start..end, project_path, filename));
-                    }
-                }
-                acp::ContentBlock::Image(_)
-                | acp::ContentBlock::Audio(_)
-                | acp::ContentBlock::Resource(_)
-                | acp::ContentBlock::ResourceLink(_) => {}
-            }
-        }
-
-        let snapshot = message_editor.update(cx, |editor, cx| {
-            editor.set_text(text, window, cx);
-            editor.buffer().read(cx).snapshot(cx)
-        });
-
-        for (range, project_path, filename) in mentions {
-            let crease_icon_path = if project_path.path.is_dir() {
-                FileIcons::get_folder_icon(false, cx)
-                    .unwrap_or_else(|| IconName::Folder.path().into())
-            } else {
-                FileIcons::get_icon(Path::new(project_path.path.as_ref()), cx)
-                    .unwrap_or_else(|| IconName::File.path().into())
-            };
-
-            let anchor = snapshot.anchor_before(range.start);
-            if let Some(project_path) = project.read(cx).absolute_path(&project_path, cx) {
-                let crease_id = crate::context_picker::insert_crease_for_mention(
-                    anchor.excerpt_id,
-                    anchor.text_anchor,
-                    range.end - range.start,
-                    filename,
-                    crease_icon_path,
-                    message_editor.clone(),
-                    window,
-                    cx,
-                );
-
-                if let Some(crease_id) = crease_id {
-                    mention_set.lock().insert(crease_id, project_path);
-                }
-            }
-        }
-
-        let snapshot = snapshot.as_singleton().unwrap().2.clone();
-        Some(snapshot.text)
     }
 
     fn handle_thread_event(
@@ -2516,34 +2238,7 @@ impl AcpThreadView {
                     .size_full()
                     .pt_1()
                     .pr_2p5()
-                    .child(div().flex_1().child({
-                        let settings = ThemeSettings::get_global(cx);
-                        let font_size = TextSize::Small
-                            .rems(cx)
-                            .to_pixels(settings.agent_font_size(cx));
-                        let line_height = settings.buffer_line_height.value() * font_size;
-
-                        let text_style = TextStyle {
-                            color: cx.theme().colors().text,
-                            font_family: settings.buffer_font.family.clone(),
-                            font_fallbacks: settings.buffer_font.fallbacks.clone(),
-                            font_features: settings.buffer_font.features.clone(),
-                            font_size: font_size.into(),
-                            line_height: line_height.into(),
-                            ..Default::default()
-                        };
-
-                        EditorElement::new(
-                            &self.message_editor,
-                            EditorStyle {
-                                background: editor_bg_color,
-                                local_player: cx.theme().players().local(),
-                                text: text_style,
-                                syntax: cx.theme().syntax().clone(),
-                                ..Default::default()
-                            },
-                        )
-                    }))
+                    .child(self.message_editor.clone())
                     .child(
                         h_flex()
                             .absolute()
@@ -2604,7 +2299,7 @@ impl AcpThreadView {
                     button.tooltip(Tooltip::text("Type a message to submit"))
                 })
                 .on_click(cx.listener(|this, _, window, cx| {
-                    this.chat(&Chat, window, cx);
+                    this.chat(window, cx);
                 }))
                 .into_any_element()
         } else {
@@ -3067,55 +2762,11 @@ impl AcpThreadView {
         paths: Vec<project::ProjectPath>,
         _added_worktrees: Vec<Entity<project::Worktree>>,
         window: &mut Window,
-        cx: &mut Context<'_, Self>,
+        cx: &mut Context<Self>,
     ) {
-        let buffer = self.message_editor.read(cx).buffer().clone();
-        let Some((&excerpt_id, _, _)) = buffer.read(cx).snapshot(cx).as_singleton() else {
-            return;
-        };
-        let Some(buffer) = buffer.read(cx).as_singleton() else {
-            return;
-        };
-        for path in paths {
-            let Some(entry) = self.project.read(cx).entry_for_path(&path, cx) else {
-                continue;
-            };
-            let Some(abs_path) = self.project.read(cx).absolute_path(&path, cx) else {
-                continue;
-            };
-
-            let anchor = buffer.update(cx, |buffer, _cx| buffer.anchor_before(buffer.len()));
-            let path_prefix = abs_path
-                .file_name()
-                .unwrap_or(path.path.as_os_str())
-                .display()
-                .to_string();
-            let completion = ContextPickerCompletionProvider::completion_for_path(
-                path,
-                &path_prefix,
-                false,
-                entry.is_dir(),
-                excerpt_id,
-                anchor..anchor,
-                self.message_editor.clone(),
-                self.mention_set.clone(),
-                self.project.clone(),
-                cx,
-            );
-
-            self.message_editor.update(cx, |message_editor, cx| {
-                message_editor.edit(
-                    [(
-                        multi_buffer::Anchor::max()..multi_buffer::Anchor::max(),
-                        completion.new_text,
-                    )],
-                    cx,
-                );
-            });
-            if let Some(confirm) = completion.confirm.clone() {
-                confirm(CompletionIntent::Complete, window, cx);
-            }
-        }
+        self.message_editor.update(cx, |message_editor, cx| {
+            message_editor.insert_dragged_files(paths, window, cx);
+        })
     }
 }
 
@@ -3132,9 +2783,6 @@ impl Render for AcpThreadView {
         v_flex()
             .size_full()
             .key_context("AcpThread")
-            .on_action(cx.listener(Self::chat))
-            .on_action(cx.listener(Self::previous_history_message))
-            .on_action(cx.listener(Self::next_history_message))
             .on_action(cx.listener(Self::open_agent_diff))
             .bg(cx.theme().colors().panel_background)
             .child(match &self.thread_state {
@@ -3430,18 +3078,17 @@ fn terminal_command_markdown_style(window: &Window, cx: &App) -> MarkdownStyle {
 }
 
 #[cfg(test)]
-mod tests {
+pub(crate) mod tests {
+    use std::{path::Path, sync::Arc};
+
     use agent_client_protocol::SessionId;
     use editor::EditorSettings;
     use fs::FakeFs;
     use futures::future::try_join_all;
     use gpui::{SemanticVersion, TestAppContext, VisualTestContext};
-    use lsp::{CompletionContext, CompletionTriggerKind};
-    use project::CompletionIntent;
+    use parking_lot::Mutex;
     use rand::Rng;
-    use serde_json::json;
     use settings::SettingsStore;
-    use util::path;
 
     use super::*;
 
@@ -3469,7 +3116,7 @@ mod tests {
         cx.deactivate_window();
 
         thread_view.update_in(cx, |thread_view, window, cx| {
-            thread_view.chat(&Chat, window, cx);
+            thread_view.chat(window, cx);
         });
 
         cx.run_until_parked();
@@ -3496,7 +3143,7 @@ mod tests {
         cx.deactivate_window();
 
         thread_view.update_in(cx, |thread_view, window, cx| {
-            thread_view.chat(&Chat, window, cx);
+            thread_view.chat(window, cx);
         });
 
         cx.run_until_parked();
@@ -3542,7 +3189,7 @@ mod tests {
         cx.deactivate_window();
 
         thread_view.update_in(cx, |thread_view, window, cx| {
-            thread_view.chat(&Chat, window, cx);
+            thread_view.chat(window, cx);
         });
 
         cx.run_until_parked();
@@ -3552,109 +3199,6 @@ mod tests {
                 .iter()
                 .any(|window| window.downcast::<AgentNotification>().is_some())
         );
-    }
-
-    #[gpui::test]
-    async fn test_crease_removal(cx: &mut TestAppContext) {
-        init_test(cx);
-
-        let fs = FakeFs::new(cx.executor());
-        fs.insert_tree("/project", json!({"file": ""})).await;
-        let project = Project::test(fs, [Path::new(path!("/project"))], cx).await;
-        let agent = StubAgentServer::default();
-        let (workspace, cx) =
-            cx.add_window_view(|window, cx| Workspace::test_new(project.clone(), window, cx));
-        let thread_view = cx.update(|window, cx| {
-            cx.new(|cx| {
-                AcpThreadView::new(
-                    Rc::new(agent),
-                    workspace.downgrade(),
-                    project,
-                    Rc::new(RefCell::new(MessageHistory::default())),
-                    1,
-                    None,
-                    window,
-                    cx,
-                )
-            })
-        });
-
-        cx.run_until_parked();
-
-        let message_editor = cx.read(|cx| thread_view.read(cx).message_editor.clone());
-        let excerpt_id = message_editor.update(cx, |editor, cx| {
-            editor
-                .buffer()
-                .read(cx)
-                .excerpt_ids()
-                .into_iter()
-                .next()
-                .unwrap()
-        });
-        let completions = message_editor.update_in(cx, |editor, window, cx| {
-            editor.set_text("Hello @", window, cx);
-            let buffer = editor.buffer().read(cx).as_singleton().unwrap();
-            let completion_provider = editor.completion_provider().unwrap();
-            completion_provider.completions(
-                excerpt_id,
-                &buffer,
-                Anchor::MAX,
-                CompletionContext {
-                    trigger_kind: CompletionTriggerKind::TRIGGER_CHARACTER,
-                    trigger_character: Some("@".into()),
-                },
-                window,
-                cx,
-            )
-        });
-        let [_, completion]: [_; 2] = completions
-            .await
-            .unwrap()
-            .into_iter()
-            .flat_map(|response| response.completions)
-            .collect::<Vec<_>>()
-            .try_into()
-            .unwrap();
-
-        message_editor.update_in(cx, |editor, window, cx| {
-            let snapshot = editor.buffer().read(cx).snapshot(cx);
-            let start = snapshot
-                .anchor_in_excerpt(excerpt_id, completion.replace_range.start)
-                .unwrap();
-            let end = snapshot
-                .anchor_in_excerpt(excerpt_id, completion.replace_range.end)
-                .unwrap();
-            editor.edit([(start..end, completion.new_text)], cx);
-            (completion.confirm.unwrap())(CompletionIntent::Complete, window, cx);
-        });
-
-        cx.run_until_parked();
-
-        // Backspace over the inserted crease (and the following space).
-        message_editor.update_in(cx, |editor, window, cx| {
-            editor.backspace(&Default::default(), window, cx);
-            editor.backspace(&Default::default(), window, cx);
-        });
-
-        thread_view.update_in(cx, |thread_view, window, cx| {
-            thread_view.chat(&Chat, window, cx);
-        });
-
-        cx.run_until_parked();
-
-        let content = thread_view.update_in(cx, |thread_view, _window, _cx| {
-            thread_view
-                .message_history
-                .borrow()
-                .items()
-                .iter()
-                .flatten()
-                .cloned()
-                .collect::<Vec<_>>()
-        });
-
-        // We don't send a resource link for the deleted crease.
-        pretty_assertions::assert_matches!(content.as_slice(), [acp::ContentBlock::Text { .. }]);
     }
 
     async fn setup_thread_view(
@@ -3673,8 +3217,6 @@ mod tests {
                     workspace.downgrade(),
                     project,
                     Rc::new(RefCell::new(MessageHistory::default())),
-                    1,
-                    None,
                     window,
                     cx,
                 )
@@ -3888,7 +3430,7 @@ mod tests {
         }
     }
 
-    fn init_test(cx: &mut TestAppContext) {
+    pub(crate) fn init_test(cx: &mut TestAppContext) {
         cx.update(|cx| {
             let settings_store = SettingsStore::test(cx);
             cx.set_global(settings_store);
