+++ crates/agent_ui/src/agent_panel.rs
@@ -11,7 +11,6 @@ use serde::{Deserialize, Serialize};
 
 use crate::NewExternalAgentThread;
 use crate::agent_diff::AgentDiffThread;
-use crate::message_editor::{MAX_EDITOR_LINES, MIN_EDITOR_LINES};
 use crate::ui::NewThreadButton;
 use crate::{
     AddContextServer, AgentDiffPane, ContinueThread, ContinueWithBurnMode,
@@ -964,8 +963,6 @@ impl AgentPanel {
                         workspace.clone(),
                         project,
                         message_history,
-                        MIN_EDITOR_LINES,
-                        Some(MAX_EDITOR_LINES),
                         window,
                         cx,
                     )
